feat: Procedural plant rendering with unique visuals (#219)
- New PlantRenderer system (src/systems/PlantRenderer.ts) extracts all
  plant rendering from GardenScene into a dedicated, decoupled system
- 22 plant species with distinct visuals across 5 growth stages
  (seed, sprout, growing, mature, wilting)
- Added strawberry, sage, watermelon to plantVisuals.ts (19→22 plants)
- Procedural idle sway for mature/growing plants with per-species intensity
- Synergy shimmer effect (alpha pulse) on synergy-active mature plants
- Seasonal color shifts via adjustSaturation from seasonalPalettes.ts
- Colorblind palette support via getActivePalette()
- Visual state cached per (plantType, growthStage, healthBucket, season)
  to avoid redundant redraws
- GardenScene reduced by ~610 lines via PlantRenderer delegation
