+++ src/systems/index.ts
@@ -26,6 +26,8 @@ export { TutorialSystem } from './TutorialSystem';
 export type { TutorialSaveData } from './TutorialSystem';
 export { AchievementSystem } from './AchievementSystem';
 export type { AchievementSaveState, AchievementState, AchievementUnlockCallback } from './AchievementSystem';
+export { PlantRenderer } from './PlantRenderer';
+export type { PlantRendererConfig } from './PlantRenderer';
 export { DailyChallengeSystem } from './DailyChallengeSystem';
 export type { DailyChallenge, LeaderboardEntry, RunHistoryEntry } from './DailyChallengeSystem';
 export { WeedSystem } from './WeedSystem';
